gimme_dat_email/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==20.1.0
email-validator==2.0.0.post2
python-whois==0.8.0
python-dotenv==1.0.0
pyhunter==1.7
